# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super STT daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/stt/super-stt-http.sock` (override with `SUPER_STT_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/stt/super-stt-http.sock\" \\\n     -H \"Authorization: Bearer $STT_TOKEN\" \\\n     http://stt.local/v1/ping\n```",
    "contact": {
      "name": "Super STT",
      "url": "https://github.com/jorge-menjivar/super-stt"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.2.4-beta.1"
  },
  "servers": [
    {
      "url": "http://stt.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/backends": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "List installed backends",
        "description": "Every backend installed on this machine, each with the models it serves, the options it exposes, and which of its secrets are configured. Secret *values* are never returned — only whether each is set.\n\nThis is the full catalog, roles included. `GET /models` is the narrower read a transcription-model picker wants; browsing what is *available to install* is `GET /registry/backends`.",
        "operationId": "list_backends",
        "responses": {
          "200": {
            "description": "The installed catalog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendCatalog"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/backends/{source}": {
      "delete": {
        "tags": [
          "backends"
        ],
        "summary": "Uninstall a backend",
        "description": "Removes an installed backend and its files from disk.\n\nIf the backend was filling a pipeline stage, that stage is emptied first: any loaded model is unloaded and the selection cleared, so the daemon does not end up pointing at files that no longer exist. The response says which stages were affected.\n\nRefused with `409 backend_busy` while a recording or realtime session is in flight — removing files out from under one would strand state it still depends on.",
        "operationId": "uninstall_backend",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "description": "The backend's `source`, percent-encoded — e.g. `github.com%2Facme%2Fwhisper`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Facme%2Fwhisper"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No installed backend claims that `source`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "409": {
            "description": "A recording or realtime session is in flight (`backend_busy`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "500": {
            "description": "The files could not be removed (`remove_failed`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/gpu_info": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Inventory the host's GPUs",
        "description": "What the daemon can see of this machine's accelerators: one entry per detected GPU with its memory, plus the host-wide driver and runtime versions that decide which backend builds will actually run here.\n\nDetection is a live probe, so this reflects the machine now rather than a cached answer. A host with no GPU answers `200` with an empty list.",
        "operationId": "get_gpu_info",
        "responses": {
          "200": {
            "description": "The detected GPUs and host toolchain versions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuInventory"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/models": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List the models the active backend can transcribe with",
        "description": "Scoped to the backend currently filling stage 1 — only its models are switchable. Post-processor models are excluded, since selecting one as a transcription model would fail every recording; the full catalog with roles is at `GET /backends`.",
        "operationId": "list_models",
        "responses": {
          "200": {
            "description": "Done.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/registry/backends": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Browse the published backend catalog",
        "description": "Every backend published to the registry, with the models each serves and whether this machine can run it. `compatible` is decided against the host's actual accelerators and this Super STT's version, so the list reflects what is installable here rather than what exists in general.\n\n`needs_client_update` separates the two ways a backend can be blocked: a host that lacks the right GPU will never run it, but a Super STT one version behind is something the user can fix in a minute. Surface those differently.\n\nThis is what is *available*; `GET /backends` is what is installed.",
        "operationId": "list_registry_backends",
        "parameters": [
          {
            "name": "include_incompatible",
            "in": "query",
            "description": "Include entries this machine cannot run. Off by default, so the catalog\nshows what is actually installable here.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by backend kind.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "online",
            "in": "query",
            "description": "Filter by whether the backend calls out to a network service.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive substring match over name and description.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The catalog could not be fetched and nothing is cached (`registry_unavailable`). Retry, or force a fetch with `POST /registry/backends/refresh`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/status": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Current model and device",
        "description": "A snapshot of what the daemon is running: which model is loaded, on which accelerator, and whether a recording cycle is in flight.\n\nSubscriber introspection and other operator detail are not exposed here — see `GET /pipeline/1` and `GET /pipeline/{stage}/model/{model}/device`, which need the `settings` scope.",
        "operationId": "status",
        "responses": {
          "200": {
            "description": "The daemon's current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonStatus"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `status` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "status"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackendCatalog": {
        "type": "object",
        "description": "Every installed backend, with its models, options and secrets.",
        "required": [
          "status",
          "backends"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackendInfo"
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "DaemonStatus": {
        "type": "object",
        "description": "What `GET /status` reports.",
        "required": [
          "status",
          "device",
          "model_loaded",
          "busy"
        ],
        "properties": {
          "busy": {
            "type": "boolean",
            "description": "`true` while a daemon-mic cycle is active — capture *and* the\ntranscription and typing that follow it. A toggle hotkey reads this and\ncalls `POST /transcribe/stop` when true, `POST /transcribe` when false."
          },
          "current_model": {
            "type": [
              "string",
              "null"
            ],
            "description": "The loaded model's name. Absent when `model_loaded` is `false`.",
            "example": "whisper-tiny"
          },
          "device": {
            "type": "string",
            "description": "The accelerator the loaded model actually runs on: `cpu`, `cuda`,\n`rocm`, `metal`, `vulkan`, or `remote` for a model served over the\nnetwork. `unknown` when nothing is loaded.",
            "example": "cuda"
          },
          "model_loaded": {
            "type": "boolean",
            "description": "`false` while the initial model is still loading, or after a failed\nswitch."
          },
          "status": {
            "type": "string",
            "description": "Always `success`.",
            "example": "success"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "recording_in_progress",
          "download_in_progress",
          "no_switch_in_progress",
          "model_not_loaded",
          "invalid_model",
          "invalid_backend",
          "cuda_unavailable",
          "invalid_device",
          "invalid_audio_theme",
          "unsupported_language",
          "invalid_value",
          "not_found",
          "internal",
          "unknown"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-stt/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "GpuInfo": {
        "type": "object",
        "description": "One GPU as reported by [`GET /gpu_info`](../../../docs/protocol/endpoints/v1/gpu_info.md).\n`vendor` is a lowercase `snake_case` tag (`nvidia` / `amd` / `intel` /\n`apple` / `unknown`). `total_bytes` is dedicated VRAM for discrete GPUs and\nthe shared system-memory ceiling for integrated/unified GPUs;\n`free_bytes` / `used_bytes` are `null` when the platform doesn't report them.",
        "required": [
          "name",
          "vendor",
          "total_bytes"
        ],
        "properties": {
          "arch_target": {
            "type": [
              "string",
              "null"
            ],
            "description": "The architecture a prebuilt asset must target to run on this GPU, in\nthe vendor's own spelling: `\"sm_86\"` on NVIDIA, `\"gfx1030\"` on AMD.\n`null` when the driver reports none (an Apple or Intel GPU, or an AMD\ncard on a kernel without KFD)."
          },
          "free_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "total_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "used_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "vendor": {
            "type": "string"
          }
        }
      },
      "GpuInventory": {
        "type": "object",
        "description": "The host's GPUs and its GPU toolchain versions.",
        "required": [
          "status",
          "gpu_info",
          "host"
        ],
        "properties": {
          "gpu_info": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GpuInfo"
            },
            "description": "One entry per detected GPU; empty on a host with none."
          },
          "host": {
            "$ref": "#/components/schemas/GpuHostInfo",
            "description": "Driver and runtime versions, independent of any one GPU."
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "ModelList": {
        "type": "object",
        "description": "The models the backend filling stage 1 can transcribe with.",
        "required": [
          "status",
          "available_models"
        ],
        "properties": {
          "available_models": {
            "type": "array",
            "items": {
              "type": "array",
              "items": false,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ]
            },
            "description": "`[name, source]` pairs. Post-processor models are excluded — they are not\nswitchable transcription models, and offering one would fail every\nrecording. The full catalog, roles included, is at `GET /backends`.",
            "example": [
              [
                "whisper-tiny",
                "github.com/super-stt/whisper"
              ]
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "RegistryError": {
        "type": "object",
        "description": "The registry surface's error envelope, which carries one extra key.\n\nThese endpoints shipped before `error_code` existed, spelling the failure\nidentity as `error`. That key is still sent, because clients read it; the\nstandard `error_code` was added alongside rather than in place of it, so the\nwhole surface honors the \"`error_code` on every error\" rule without breaking\nanyone. Both name the same failure and always agree.\n\nDocumented as its own shape rather than folded into [`ErrorEnvelope`]:\n`error` appears *only* here, and putting it on the shared envelope would\ntell every other endpoint's reader to expect a key they will never receive.",
        "required": [
          "status",
          "error_code",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "The same identifier under the key this surface has always used. Retained\nfor clients written against it; prefer `error_code`.",
            "example": "not_found"
          },
          "error_code": {
            "type": "string",
            "description": "The stable identifier for this failure.",
            "example": "not_found"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail, when there is any to add."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
      "RegistryListResponse": {
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "backends"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistryBackend"
            }
          },
          "generated_at": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "UninstallResponse": {
        "type": "object",
        "required": [
          "uninstalled",
          "was_active"
        ],
        "properties": {
          "uninstalled": {
            "type": "boolean"
          },
          "was_active": {
            "type": "boolean",
            "description": "The backend was filling stage 1, which was emptied before the files\nwent."
          },
          "was_post_processor": {
            "type": "boolean",
            "description": "The backend was filling stage 2 — selected as the post-processor\nbackend, loaded or not — which was emptied before the files went.\nAbsent from an older daemon's answer, which reads as `false`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}